requirements.txt
azure-core==1.30.0
azure-storage-blob==12.19.0
beautifulsoup4==4.12.3
pymongo==4.6.1
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.1
selenium==4.18.1